Added role based Authentication

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.SQLOutput;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -27,8 +28,7 @@ public class JournalEntryService {
          journalEntry.setDate(LocalDateTime.now());
          JournalEntry saved = journalEntryRepository.save(journalEntry);
          user.getJournalEntries().add(saved);
-
-         userService.saveEntry(user);
+         userService.saveUser(user);
      } catch (Exception e) {
          System.out.println(e);
          throw new RuntimeException("An error occurred while saving the entry" , e);
@@ -36,25 +36,38 @@ public class JournalEntryService {
    }
 
     public void saveEntry(JournalEntry journalEntry){
-
         journalEntryRepository.save(journalEntry);
     }
 
     public List<JournalEntry>getAll(){
-
        return journalEntryRepository.findAll();
    }
 
    public Optional<JournalEntry> findById(ObjectId id){
-
        return journalEntryRepository.findById(id);
    }
 
-   public void deleteById(ObjectId id, String userName){
-       User user = userService.findByUsername(userName);
-       user.getJournalEntries().removeIf(x-> x.getId().equals(id));
-       userService.saveEntry(user);
-         journalEntryRepository.deleteById(id);
+
+   @Transactional
+   public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+       try{
+           User user = userService.findByUsername(userName);
+             removed =  user.getJournalEntries().removeIf(x-> x.getId().equals(id));
+           if(removed){
+               userService.saveNewUser(user);
+               journalEntryRepository.deleteById(id);
+           }
+       }catch(Exception e){
+           System.out.println(e);
+           throw new RuntimeException("An error occurred while deleting the entry" , e);
+       }
+    return removed;
    }
+
+//   public List<JournalEntry> findByUserName(String userName){
+////       User user = userService.findByUsername(userName);
+///      return user.getJournalEntries();
+//   }
 
 }
